Kayttajatiedot ja sen kaverityypit omaan filuun
