+++ server/src/main/kotlin/fi/oph/ludos/auth/OppijanumerorekisteriClient.kt
@@ -11,15 +11,15 @@ import java.util.*
 class OppijanumerorekisteriClient : CasAuthenticationClient("oppijanumerorekisteri-service") {
     private val logger = LoggerFactory.getLogger(javaClass)
 
-    fun getUserDetailsByOid(oid: String): OppijanumeroRekisteriHenkilo? {
+    fun getUserDetailsByOid(oid: String): OppijanumerorekisteriHenkilo? {
         val req = OphHttpRequest.Builder.get("https://$opintopolkuHostname/oppijanumerorekisteri-service/henkilo/$oid")
             .build()
         return try {
-            val response = httpClient.execute<OppijanumeroRekisteriHenkilo>(req)
+            val response = httpClient.execute<OppijanumerorekisteriHenkilo>(req)
             response.handleErrorStatus(404).with { Optional.empty() }
             response
                 .expectedStatus(HttpStatus.OK.value())
-                .mapWith { s -> mapper.readValue<OppijanumeroRekisteriHenkilo>(s) }
+                .mapWith { s -> mapper.readValue<OppijanumerorekisteriHenkilo>(s) }
                 .orElse(null)
         } catch (e: Exception) {
             logger.warn("Unexpected error getting oppijanumerorekisteri details for $oid", e)
@@ -27,3 +27,24 @@ class OppijanumerorekisteriClient : CasAuthenticationClient("oppijanumerorekiste
         }
     }
 }
+
+data class OppijanumerorekisteriHenkilo(
+    val etunimet: String,
+    val kutsumanimi: String,
+    val sukunimi: String,
+    val asiointiKieli: Asiointikieli?
+) {
+    constructor(kayttajatiedot: Kayttajatiedot) : this(
+        kayttajatiedot.etunimet ?: "-",
+        kayttajatiedot.etunimet ?: "",
+        kayttajatiedot.sukunimi ?: "-",
+        null
+    )
+
+    fun formatName(): String {
+        val etu = kutsumanimi.ifBlank { etunimet }
+        return "$etu $sukunimi"
+    }
+}
+
+data class Asiointikieli(val kieliKoodi: String, val kieliTyyppi: String)
